requirements.backend.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart
pydantic
email-validator
pandas
openpyxl